[IMP] sale_planner_calendar: Merge files calendar.py and sale_planner_calendar_event.py in calendar_event.py
