+++ sale_planner_calendar/models/__init__.py
@@ -1,5 +1,5 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-from . import calendar
+from . import calendar_event
 from . import calendar_event_type
 from . import calendar_attendee
 from . import res_config_settings
@@ -7,6 +7,5 @@ from . import res_partner
 from . import sale_order
 from . import sale_payment_sheet
 from . import sale_planner_calendar_event_profile
-from . import sale_planner_calendar_event
 from . import sale_planner_calendar_issue_type
 from . import sale_planner_calendar_summary
